04  23022024_sesion04

+++ src/controllers/cart.controllers.js
@@ -1,15 +1,23 @@
 const catchError = require('../utils/catchError');
 const Cart = require('../models/Cart');
 const Product = require('../models/Product');
+const Category = require('../models/Category');
 
 const getAll = catchError(async (req, res) => {
   const userId = req.user.id
   const results = await Cart.findAll({
     where:{userId},
-    include:{
+    include:[
+      // Product
+      {
       model: Product,
-      attributes:{exclude:["updatedAt", "createdAt"]}
+      attributes:{exclude:["updatedAt", "createdAt"]},
+      include: {
+        model: Category,
+        attributes:['name']
+      }
     }
+  ]
   });
   return res.json(results);
 });
@@ -26,15 +34,18 @@ const create = catchError(async (req, res) => {
 
 const remove = catchError(async (req, res) => {
   const { id } = req.params;
-  await Cart.destroy({ where: { id } });
+  const userId = req.user.id
+  const result = await Cart.destroy({ where: { id, userId } });
+  if(!result) return res.sendStatus(404)
   return res.sendStatus(204);
 });
 
 const update = catchError(async (req, res) => {
+  const userLogged = req.user.id
   const { id } = req.params;
   const result = await Cart.update(
-    req.body,
-    { where: { id }, returning: true }
+    {quantity},
+    { where: { id, userId: userLogged }, returning: true }
   );
   if (result[0] === 0) return res.sendStatus(404);
   return res.json(result[1][0]);
